requirements.txt
astroid==2.3.3
biopython==1.75
cycler==0.10.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pkg-resources==0.0.0
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
rope==0.14.0
scipy==1.3.2
seaborn==0.9.0
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
